+++ node/rpc-api/src/lib.rs
@@ -90,6 +90,9 @@ pub static ACCESS_MAP: Lazy<HashMap<&str, Access>> = Lazy::new(|| {
     access.insert(net_api::NET_CONNECT, Access::Write);
     access.insert(net_api::NET_DISCONNECT, Access::Write);
 
+    // DB API
+    access.insert(db_api::DB_GC, Access::Write);
+
     access
 });
 
@@ -415,4 +418,11 @@ pub mod net_api {
     pub const NET_DISCONNECT: &str = "Filecoin.NetDisconnect";
     pub type NetDisconnectParams = (String,);
     pub type NetDisconnectResult = ();
+}
+
+/// DB API
+pub mod db_api {
+    pub const DB_GC: &str = "Filecoin.DatabaseGarbageCollection";
+    pub type DBGCParams = ();
+    pub type DBGCResult = ();
 }