+++ node/rpc/src/sync_api.rs
@@ -153,6 +153,8 @@ mod tests {
         };
         let (new_mined_block_tx, _) = flume::bounded(5);
         let start_time = OffsetDateTime::now_utc();
+        let (gc_event_tx, _) = flume::unbounded();
+
         let state = Arc::new(RPCState {
             state_manager,
             keystore: Arc::new(RwLock::new(KeyStore::new(KeyStoreConfig::Memory).unwrap())),
@@ -165,6 +167,7 @@ mod tests {
             chain_store: cs_for_chain,
             beacon,
             new_mined_block_tx,
+            gc_event_tx,
         });
         (state, network_rx)
     }